Panic on error in main context
In the main context we need to panic on error, as we may not yet be
listening on the <-ctx.quit channel.

In the timeout goroutine, we need to interrupt the otto context, and
send the error to the <-ctx.quit channel so that we can quit the main
context, and quit any asynchronous code at the same time.

There is now no need to mark the orbit context as ‘done’ as panicing on
error will end the context anyway.